refactor: 헤더와 map 컴포넌트 분리 (#88)
